Added enemy healthbars
This was something I did after I completed the game in 2014, but then lost it. Added healthbars overtop of enemies, and static healthbars for bosses. Also fixed a weird bug I can't believe I didn't notice before, so now it only draws objects if they are within the screen (which is what I meant for it to do years back)