Sword and boss updates, added hearts
Reorganized to have the sword as part of the character class. Also added in the chance of a few hearts upon defeat of an enemy if the character's health is low. Finally, added more brains to the boss, now it will try to run from a sword swing if it's health is low.