vertical motion update
When the character jumps, the level no longer moves up and down, but the character does. When he lands, the level then centers itself around the character, so that the player can see a larger portion of  the level. However, there is a glitch with the power items that needs to be fixed, the hitbox somehow got extended